third_party/blink/web_tests/platform/mac-mac11-arm64/http/tests/security/video-poster-cross-origin-crash-expected.txt

CONSOLE ERROR: Access to image at 'http://localhost:8080/misc/resources/compass.jpg' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
>>>
